cardiovascular_autonomic_domain/requirements.txt
numpy
pyserial
PySide6
opencv-python
